feat: add web fallback chat modal and polish for Ava Anywhere (M4)
Extract ChatOverlayContent as shared component used by both the Electron
overlay window and the new web fallback Dialog modal. Add Cmd+K / Ctrl+K
global keyboard shortcut to toggle the chat modal in web mode.

Add slide-in animation for conversation history panel. Add chatModalOpen
state to the chat store.
